Update and rename Table_S2.csv to Table_S2-Output of the sequencing and the bioinformatics analysis.csv
The names of the libraries are provided in the "pool" column. The library names were generated to include the main variables of the study: mosquito species (Culex: Cx, Aedes: Ae), sampling zone, year, sex (female: F, male: M) and replicate (number of the replicate library from the same species, zone, year and sex). The "number" column provides the number of mosquitoes pooled in the library. The remaining columns represent the output of the bioinformatics pipeline, with the number of total reads (“Total_read”), the number of reads with a viral hit (“Reads_with_viral_hit”), the porportion of viral reads (%_viral_reads), the number of reads that mapped on insecta (“Map_on_Insecta”) or insect (“Map_on_Bacteria”), the number of contigs (“Nb_of_contigs”), the information on the size of the contigs (“Min_contigs_length”, “Max_contigs_length”, “Avg_contigs_length”) and the number of contigs with a viral hit (“Contigs_with_viral_hit”).