Update and rename Table_S4.csv to Table_S4-Output statistics of Generalized Linear Mixed models.csv
The upper table displays the model output using the “glmmTMB” function implemented in R. The lower table presents the results of pairwise comparisons, obtained using the emmeans function. The following significance codes are used to indicate the level of statistical significance: 0, '***', 0.001, '**', 0.01, '*', 0.05, '.' and 1. For each table, the models were run on either a dataset containing all libraries or subsets of libraries of each mosquito species (Culex quinquefasciatus or Aedes aegypti).