Update and rename Table_S5.csv to Table_S5-Output statistics of the PERMANOVA models.csv
The table on the top shows the results for the complete dataset. The second and third tables show the results for a subset of the data only including libraries of Culex quinquefasciatus or Aedes aegypti, respectively. The following significance codes are used to indicate the level of statistical significance: 0, '***', 0.001, '**', 0.01, '*', 0.05, '.' and 1.