Update and rename Table_S6.csv to Table_S6-Output statistics of the IndVal function and data used in defining new viral species.csv
Each indicator VTU of one of the two sexes in Culex (top table) and Aedes (bottom table) is listed here (column “VTU”) with the statistic (column “stat”) and the associated p-value (column “p-value”). The following significance codes are used to indicate the level of statistical significance: 0, '***', 0.001, '**', 0.01, '*', 0.05, '.' and 1. The species name (or new species name for new virus species - pident < 90% at amino acid level) is also given (column “Species name / New name”), together with the statistics for the length of the best contig. The column designated as "qlen" represents the length of the query sequence, the column labeled "slen" represents the length of the subject sequence, while the column identified as "length" represents the length of the alignment (sequence overlap).