Update and rename Table_S7.csv to Table_S7-Output statistics from the R “binGroup” package.csv
All VTUs are listed in the table in the column designated "VTU," along with the calculation of the infection rate ("p") by mosquito species ("species”) and the sex of the host mosquito ("sex"). The values of the lower confidence limit and the upper confidence limit are displayed in the columns designated "lcl" and "ucl," respectively.