Table_S3-Output of the homology and taxonomy search for the Aedes dataset
The "Best-hit" column provides the accession with the lowest e-value found by Diamond. The “VTU” column contains the VTU names. The “cluster” to “genus” columns provide the VTU taxonomy as stated by the ICTV. All columns from "sum_reads" to "max coverage" provide information on the contigs of each VTU (column fields described in the legend of Table_S2).