Table S1: Sampling sites of the study.
The name of the site is given in the “Site” column. The EID agency responsible for the captures is indicated in the “Agency” column. The other columns describe the site in the following order (from left to right): the distance to the first dwelling, the town in which the site is located, the number of inhabitants of the town in 2019, the latitude and the longitude. The “Urban”, “Cropland” and “Natural_Semi-natural” columns were used for classifications according to the level of urbanisation. The values for each site represent the number of “pixels” of 1 km2 belonging to the group in question. Depending on the proportion of each group (see Materials and Methods), the sites were classified as urban or rural, indicated in the “Urbanization_level” column.