Table S3: Read count by VTU and taxonomy.
The "VTU" column provides the name given to the VTU. The "Best-hit" column provides the accession with the lowest e-value found by Diamond. The "Nucleic_acid" and "genome" columns provide information on the type of VTU genome. The "cluster" column was created to add a taxonomic rank similar to a family, in order to avoid the lack of taxonomic information for certain viruses. The "clade" to "genus" columns provide the VTU taxonomy as stated by the ICTV. All columns from "qlen" to "bitscore" provide information on the contigs of each VTU and are Blastn output. All other columns represent count data by libraries for each VTU. The "sum_reads" column provides information on the total number of reads detected for the VTU (across all libraries).