Table S4: Output statistics of Generalized Linear Mixed models.
The upper table shows the results of the model using the “glmmTMB” function implemented in R. The columns represent the statistical results for each variable tested: “Urbanization”, “Month” and “Year”. Each row represents a diversity index: “Species richness”, “Shannon” and “Simpson”.  
The lower table shows the results of pairwise comparisons, obtained using the "emmeans" function for each diversity index (in columns). The following significance codes are used to indicate the level of statistical significance: 0, '***', 0.001, '**', 0.01, '*', 0.05, '.' and 1.